requirements.txt
requests
beautifulsoup4
lxml
schedule
python-dotenv